requirements.txt
streamlit==1.46.1
opencv-python==4.10.0
pillow==11.1.0
numpy==1.26.4 